Revert "Removing the constraint of unnecessary mutability for the Range<T> components."
This reverts commit 9eb357e5e7e3c4534261dca95ae999bd663aad5e.

The commit introduces a bug where the whole Range<T> object would also become implicitly constant if its members are. That lead to undefined behavior when assigning to old Range<T> objects. At least GNU g++ was unable to catch the issue at compile time, and it manifested in the form of a straight segmentation fault.